adiciona a camada de services e separa as entidades de usuario e jogos
