+++ app/services/jogo_services.py
@@ -1,0 +1,24 @@
+import requests
+
+from ..views.jogo_view import serializar_jogos
+from ..utils import auth
+from .. import config
+
+@auth.token_required
+def listar_jogos_steam(contexto):
+    usuario = contexto["usuario"]
+    if not usuario.steam_id:
+        return {"erro": "nenhuma conta da steam associada a este usuario"}, 400
+    headers = {"x-webapi-key": config.STEAM_API_KEY}
+    parametros = {
+      "steamid": usuario.steam_id,
+      "include_appinfo": True,
+      "include_played_free_games": True
+    }
+    resposta = requests.get(
+      'https://api.steampowered.com/IPlayerService/GetOwnedGames/v1/',
+      headers=headers,
+      params=parametros,
+      timeout=15
+    )
+    return serializar_jogos(resposta.json()), 200
