+++ app/views/usuario_view.py
@@ -5,14 +5,3 @@ def serializar_usuario(usuario):
       "email": usuario.email,
       "data_nascimento": usuario.data_nascimento.isoformat()
     }
-
-def serializar_jogos(resposta):
-    quantidade_jogos = resposta["response"]["game_count"]
-    jogos = []
-    for game in resposta["response"]["games"]:
-        jogos.append(game["name"])
-    jogos.sort()
-    return {
-      "quantidade_jogos": quantidade_jogos,
-      "jogos": jogos
-    }
